nghbrly/requirements.txt
asgiref==3.6.0
Django==2.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
PyJWT==2.6.0
python-dotenv==0.21.1
pytz==2022.7.1
sqlparse==0.4.3
